server/requirement.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart  # for file uploads

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface

# Vectorstore
# chromadb
pinecone

# Embeddings
# Needed for HF sentence embedding models
sentence-transformers

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Logging (optional but recommended)
loguru



#  uv pip install -r requirements.txt